hnsw_optimization/requirements.txt
hnswlib>=0.7.0
faiss-cpu>=1.7.4
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
joblib>=1.1.0
tqdm>=4.62.0
matplotlib>=3.5.0
